+++ convoy/settings.py
@@ -92,6 +92,13 @@ BatchShipyardSettings = collections.namedtuple(
         'store_timing_metrics',
     ]
 )
+DockerRegistrySettings = collections.namedtuple(
+    'DockerRegistrySettings', [
+        'allow_public_docker_hub_pull_on_missing',
+        'storage_account', 'container', 'server', 'port',
+        'user', 'password',
+    ]
+)
 DataReplicationSettings = collections.namedtuple(
     'DataReplicationSettings', [
         'peer_to_peer', 'non_peer_to_peer_concurrent_downloading',
@@ -498,16 +505,17 @@ def credentials_storage(config, ssel):
     )
 
 
-def docker_registry_hub_login(config):
-    # type: (dict) -> tuple
-    """Get docker registry hub login settings
+def docker_registry_login(config, server):
+    # type: (dict, str) -> tuple
+    """Get docker registry login settings
     :param dict config: configuration object
+    :param str server: credentials for login server to retrieve
     :rtype: tuple
     :return: (user, pw)
     """
     try:
-        user = config['credentials']['docker_registry']['hub']['username']
-        pw = config['credentials']['docker_registry']['hub']['password']
+        user = config['credentials']['docker_registry'][server]['username']
+        pw = config['credentials']['docker_registry'][server]['password']
         if util.is_none_or_empty(user) or util.is_none_or_empty(pw):
             raise KeyError()
     except KeyError:
@@ -654,28 +662,41 @@ def batch_shipyard_encryption_public_key_pem(config):
     return pem
 
 
-def docker_registry_private_allow_public_pull(config):
-    # type: (dict) -> bool
-    """Get public docker hub pull on missing setting
+def docker_registry_private_settings(config):
+    # type: (dict) -> DockerRegistrySettings
+    """Get docker private registry backed to azure storage settings
     :param dict config: configuration object
-    :rtype: bool
-    :return: if public pull is allowed on missing
+    :rtype: DockerRegistrySettings
+    :return: docker registry settings
     """
     try:
         pregpubpull = config['docker_registry']['private'][
             'allow_public_docker_hub_pull_on_missing']
     except KeyError:
         pregpubpull = False
-    return pregpubpull
-
-
-def docker_registry_azure_storage(config):
-    # type: (dict) -> tuple
-    """Get docker private registry backed to azure storage settings
-    :param dict config: configuration object
-    :rtype: tuple
-    :return: (storage account name, container)
-    """
+    try:
+        server = config['docker_registry']['private']['server']
+        if util.is_none_or_empty(server):
+            raise KeyError()
+        server = server.split(':')
+        if len(server) == 1:
+            port = 80
+        elif len(server) == 2:
+            port = int(server[1])
+        else:
+            raise ValueError('invalid docker registry server specification')
+        server = server[0]
+        # get login
+        user, pw = docker_registry_login(config, server)
+        if util.is_none_or_empty(user) or util.is_none_or_empty(pw):
+            raise ValueError(
+                'Docker registry login settings not specified for: {}'.format(
+                    server))
+    except KeyError:
+        server = None
+        port = None
+        user = None
+        pw = None
     try:
         sa = config['docker_registry']['private']['azure_storage'][
             'storage_account_settings']
@@ -688,7 +709,19 @@ def docker_registry_azure_storage(config):
     except KeyError:
         sa = None
         cont = None
-    return sa, cont
+    if server is not None and sa is not None:
+        raise ValueError(
+            'cannot specify both a private registry server host and a '
+            'private registry backed by Azure Storage')
+    return DockerRegistrySettings(
+        allow_public_docker_hub_pull_on_missing=pregpubpull,
+        storage_account=sa,
+        container=cont,
+        server=server,
+        port=port,
+        user=user,
+        password=pw,
+    )
 
 
 def data_replication_settings(config):